requirements.txt
arcgis>=1.8.1
dask>=2.20.0
python-dotenv>=0.14.0
requests>=2.24.0
shapely>=1.7.0
swifter>=0.305
tqdm>=4.47.0
